Using GGG: use a unordered list for GGG-specific PR guidance
In a subsequent commit we want to mention that it is prefered to use an
empty desccription for a single commit PR, since the description is
copied as an in-patch commentary and having the same text twice (once in
the commit messge and once in the in-patch commentary) decreases
the signal-to-noise ratio for reviewers.

Before doing so, let's convert part of the first paragraph of the "How
can you use GitGitGadget?" section into an unorderded list. This will
make the text clearer when we mention single commit PRs later.