Using GGG: instruct to drop description for single-commit PRs
For single commit PRs, GitHub prepends the commit message to the PR
description, before the PR template.

Contributors will then (hopefully) remove the template and keep the rest
of the description. Since the PR description is added as an in-patch
commentary for single-commit PRs, the same text appears twice in the
generatred email (once in the commit messge and once in the in-patch
commentary), decreasing the signal-to-noise ratio for reviewers.

Let's instruct readers to empty the PR description for single commit
PRs.

This partly addresses:
https://github.com/gitgitgadget/gitgitgadget/issues/340