Don't specify the password on the command line
https://www.netmeister.org/blog/passing-passwords.html has a decent
overview of the security issues with that.

Since this is interactive, getting curl to prompt for the password is
probably the easiest alternative. From
https://curl.se/docs/manpage.html#-u

> If you simply specify the username, curl prompts for a password.

I don't have an easy way to test the actual command in this doc, but the
flag does seem to behave as documented at least:

```
$ curl --user user@example.com https://google.com
Enter host password for user 'user@example.com'
```

Fixes gitgitgadget/gitgitgadget#1843