PR build: when called from forks, publish a Git bundle
When the HTML page(s) are not up to date, the PR build re-generates
them, commits the result and wants to push it to the branch.

However, this only works if the PR branch and the target branch are in
the same repository. Otherwise the workflow run's `GITHUB_TOKEN` lacks
write permission, even if the PR author gave permission to the
maintainers to push to that branch.

Let's fall back to doing the next best thing and publish a Git bundle,
and provide instructions how to fetch from them and update the PR with
those changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>