Drop the current PR automation
As I found out the hard way, it is quite the challenge to keep the HTML
pages up to date when the source of truth is in newly-added Markdown
files.

The simple reason for this is that generated files should not need to be
committed. Instead, I would like to use Hugo (https://gohugo.io/) to
render the HTML pages on the fly while deploying to GitHub Pages.

Side note: The paved path on GitHub is to use Jekyll, see
https://docs.github.com/en/pages/setting-up-a-github-pages-site-with-jekyll/about-github-pages-and-jekyll

However, I found out that Jekyll is really, really slow (probably
because Ruby is really, really slow) and I do not want to waste
resources like that. If the migration of https://git-scm.com/ to a
static site generator is any indicator, we are talking about 2-3 orders
of magnitude here, so it's really no laughing matter.

Unfortunately, since GitHub sees unfit to support a better paved path,
this means that we will have to cobble together advice from
https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site#creating-a-custom-github-actions-workflow-to-publish-your-site
and from https://gohugo.io/host-and-deploy/host-on-github-pages/ to use
this, but the good news is that I went through all of those steps in the
migration of https://git-scm.com/ and can put that experience to good
use here, too.

So here is the first step: drop the automation that tries to support the
original way to build GitGitGadget's home page.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>